Fix RTLSDR and adjustments dialogs.

+++ src/SettingsDialog.py
@@ -224,30 +224,47 @@ class SettingsDialog:
         
     def rmRpt(self, *args):
         row = self.repolist.get_selected_row()
-        for item in self.config['Repeaters']:
-            if self.config['Repeaters'][item] == row.url:
-                del self.config['Repeaters'][item]
-                #TODO clean up files
-        self.repolist.remove( row )
+        if row:
+            for item in self.config['Repeaters']:
+                if self.config['Repeaters'][item] == row.url:
+                    del self.config['Repeaters'][item]
+                    #TODO clean up files
+            self.repolist.remove( row )
+        else:
+            self.msgNoneSelected()
         
     def propertyRpt(self, *args):
-        url = self.doDialog('Re/set current selected url:',self.repolist.get_selected_row().url)
-        tmpfile = userFile('repeater-temp-file.csv')
-        if url and url.find('.csv') > -1 :
-            #verify and get the name:
-            try:
-                #TODO should be in a background thread somehow
-                urllib.request.urlretrieve(url, tmpfile)
-                csv = CsvRepeaterListing(tmpfile)
-                if csv.name:
-                    self.config['Repeaters'][csv.name] = url
-                    self.repolist.get_selected_row().url = url
-                    self.repolist.show_all()
-                else:
-                    self.doMsg('Error:','Invalid entry? Must be a .csv file available on a server.')
-                os.remove(tmpfile)
-            except HTTPError:
-                self.doMsg('Error:','HTTP error- Must be a .csv file available on a server.')
+        if self.repolist.get_selected_row():
+            url = self.doDialog('Re/set current selected url:',self.repolist.get_selected_row().url)
+            tmpfile = userFile('repeater-temp-file.csv')
+            if url and url.find('.csv') > -1 :
+                #verify and get the name:
+                try:
+                    #TODO should be in a background thread somehow
+                    urllib.request.urlretrieve(url, tmpfile)
+                    csv = CsvRepeaterListing(tmpfile)
+                    if csv.name:
+                        self.config['Repeaters'][csv.name] = url
+                        self.repolist.get_selected_row().url = url
+                        self.repolist.show_all()
+                    else:
+                        self.doMsg('Error:','Invalid entry? Must be a .csv file available on a server.')
+                    os.remove(tmpfile)
+                except HTTPError:
+                    self.doMsg('Error:','HTTP error- Must be a .csv file available on a server.')
+        else:
+            self.msgNoneSelected()
+    
+    def msgNoneSelected(self):
+        dialogWindow = Gtk.MessageDialog(self.dialog,
+            Gtk.DialogFlags.MODAL | Gtk.DialogFlags.DESTROY_WITH_PARENT,
+            Gtk.MessageType.ERROR,
+            Gtk.ButtonsType.OK,
+            'Select an entry first.')
+        dialogWindow.set_title('Choose a source')
+        dialogWindow.show_all()
+        response = dialogWindow.run()
+        dialogWindow.destroy()
             
 
     def onDestroy(self, *args):
